Optimize passing CPointer to nativecall ever so slightly
We emit an unbox_i operation in a compiled nativecall body in order to
get the pointer value to pass to the native function. Without a spesh
method on the CPointer repr, this would interpret as, and jit into, a
call to CPointer's get_int.

Instead of a call, we just emit a spesh op to do the memory offset and
read for us, which the JIT also likes very much.