Use libc allocator functions for NativeCall
Because the memory could be passed to a module and freed there, we can't
use the `MVM_*` functions because those could be using a different
allocator (e.g., mimalloc as is the new default build configuration). In
cases where we can't specifiy the function to use, we instead explicitly
allocate another chunk of memory with libc functions, copy what we
originally got into it, then free the original version.